// user-service/src/common/config/custom-config.enum.ts
export enum EnvKey {
  NODE_ENV = 'NODE_ENV',
  NODE_PORT = 'NODE_PORT',

  DATABASE_HOST = 'DATABASE_HOST',
  DATABASE_PORT = 'DATABASE_PORT',
  DATABASE_USER = 'DATABASE_USER',
  DATABASE_PASSWORD = 'DATABASE_PASSWORD',
  DATABASE_NAME = 'DATABASE_NAME',

  RABBIT_MQ_URI = 'RABBIT_MQ_URI',
  RABBIT_MQ_TIMEOUT = 'RABBIT_MQ_TIMEOUT',

  JWT_SECRET = 'JWT_SECRET',
  JWT_TOKEN_LIFETIME_REMEMBER_ME = 'JWT_TOKEN_LIFETIME_REMEMBER_ME',
  JWT_TOKEN_LIFETIME_NORMAL = 'JWT_TOKEN_LIFETIME_NORMAL',

  AUTH_COOKIE_NAME = 'AUTH_COOKIE_NAME',
  AUTH_COOKIE_DOMAIN = 'AUTH_COOKIE_DOMAIN',

  LANDING_URL = 'LANDING_URL',
  RESEND_EMAIL_WAIT_MINUTES = 'RESEND_EMAIL_WAIT_MINUTES',

  ALLOWED_ORIGINS = 'ALLOWED_ORIGINS',
}